# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LemonLens API",
        "description": "Vehicle complaint and report management system API",
        "contact": {
            "name": "LemonLens Support",
            "email": "support@lemonlens.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "LemonLens API Server"
        }
    ],
    "paths": {
...
        "/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get authenticated user's profile",
                "description": "Get authenticated user's profile",
                "operationId": "f41c2f0174dabc7c31e42096a2b24459",
                "responses": {
                    "200": {
                        "description": "User profile information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email"
                                        },
                                        "credits": {
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/me/credits/history": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get authenticated user's credit transaction history",
                "description": "Get user's credit transaction history",
                "operationId": "84ea799935bf835bd1ec1121a37887fa",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of transactions to return (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Credit transaction history",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_credits": {
                                            "type": "integer"
                                        },
                                        "transactions": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "amount": {
                                                        "type": "integer"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "addition",
                                                            "deduction"
                                                        ]
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "balance_before": {
                                                        "type": "integer"
                                                    },
                                                    "balance_after": {
                                                        "type": "integer"
                                                    },
                                                    "metadata": {
                                                        "type": "object"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/me/credits/stats": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get authenticated user's credit usage statistics",
                "description": "Get user's credit statistics",
                "operationId": "3f0b4b3d5f2d1b92018d4d4b5ffc1b3e",
                "responses": {
                    "200": {
                        "description": "Credit usage statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_credits": {
                                            "type": "integer"
                                        },
                                        "total_credits_used": {
                                            "type": "integer"
                                        },
                                        "total_credits_used_this_month": {
                                            "type": "integer"
                                        },
                                        "reports_generated": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}